requirements.txt
# Function dependencies, for example:
google-cloud-bigquery==1.27.2
google-cloud-dlp==3.0.1
pandas==1.2.3
neo4j==4.2.1
